+++ src/Git_Updater/API/GitHub_API.php
@@ -195,19 +195,18 @@ class GitHub_API extends API implements API_Interface {
 	 *
 	 * @param array  $response HTTP headers.
 	 * @param string $repo     Repo name.
+	 *
+	 * @return void|int
 	 */
 	public static function ratelimit_reset( $response, $repo ) {
 		if ( isset( $response['headers']['x-ratelimit-reset'] ) ) {
 			$reset = (int) $response['headers']['x-ratelimit-reset'];
 			//phpcs:ignore WordPress.DateTime.RestrictedFunctions.date_date
 			$wait                        = date( 'i', $reset - time() );
-			static::$error_code[ $repo ] = array_merge(
-				static::$error_code[ $repo ],
-				[
-					'git'  => 'github',
-					'wait' => $wait,
-				]
-			);
+			static::$error_code[ $repo ] = isset( static::$error_code[ $repo ] ) ? static::$error_code[ $repo ] : [];
+			static::$error_code[ $repo ] = array_merge( static::$error_code[ $repo ], [ 'wait' => $wait ] );
+
+			return $wait;
 		}
 	}